Try to make linear layer even faster

+++ source/bench/linear/source/main.cpp
@@ -12,13 +12,13 @@ using namespace qconv::testutils;
 
 void linear_naive(benchmark::State& state)
 {
-  alignas(Alignment) int8_t input[512];
-  alignas(Alignment) int8_t weights[32][512];
+  alignas(Alignment) int32_t input[512];
+  alignas(Alignment) int32_t weights[32][512];
   alignas(Alignment) int32_t biases[32];
 
   modInit(input, 512, 11);
   modInit(biases, 32, 11);
-  modInit<int8_t, 32, 512>(weights, 128);
+  modInit<int32_t, 32, 512>(weights, 128);
   LinearNaive<512, 32> l;
   l.init(weights, biases);
   for (auto _ : state) {
@@ -34,13 +34,13 @@ BENCHMARK(linear_naive);
 
 void linear_simdops(benchmark::State& state)
 {
-  alignas(Alignment) int8_t input[512];
-  alignas(Alignment) int8_t weights[32][512];
+  alignas(Alignment) int32_t input[512];
+  alignas(Alignment) int32_t weights[32][512];
   alignas(Alignment) int32_t biases[32];
 
   modInit(input, 512, 11);
   modInit(biases, 32, 11);
-  modInit<int8_t, 32, 512>(weights, 128);
+  modInit<int32_t, 32, 512>(weights, 128);
   Linear<512, 32> l;
   l.init(weights, biases);
   for (auto _ : state) {
